console: Refactor application title section
